todoapp/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.6.1
pytest-django==4.8.0
